# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - temporary email service with custom subdomains and generated addresses
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subdomain {
  id: number;
  name: string;
  /** @nullable */
  webhookSecret?: string | null;
  createdAt: string;
  emailCount: number;
  unreadCount: number;
}

export interface SubdomainInput {
  /**
   * @minLength 1
   * @maxLength 253
   */
  name: string;
  webhookSecret?: string;
}

export interface TempAddress {
  id: number;
  /** The part before the @ symbol */
  localPart: string;
  /** The domain part (matches a registered subdomain name) */
  domain: string;
  /** Full email address (localPart@domain) */
  fullAddress: string;
  subdomainId: number;
  subdomainName: string;
  createdAt: string;
  emailCount: number;
  unreadCount: number;
}

export interface AddressInput {
  /** Which subdomain domain to generate an address for */
  subdomainId: number;
  /** Optional custom local part; auto-generated if omitted */
  localPart?: string;
}

export interface Email {
  id: number;
  subdomainId: number;
  subdomainName: string;
  /**
   * The generated address this email was matched to, if any
   * @nullable
   */
  addressId?: number | null;
  fromAddress: string;
  toAddress: string;
  subject: string;
  bodyText: string;
  /** @nullable */
  bodyHtml?: string | null;
  receivedAt: string;
  isRead: boolean;
}

export interface WebhookEmailPayload {
  from: string;
  to: string;
  subject: string;
  bodyText: string;
  bodyHtml?: string;
}

export interface StatsSummary {
  totalSubdomains: number;
  totalAddresses: number;
  totalEmails: number;
  unreadEmails: number;
  emailsToday: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListRecentEmailsParams = {
  limit?: number;
};