# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TempMail API - temporary email service with custom subdomains and generated addresses
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subdomains
    description: Subdomain management
  - name: addresses
    description: Generated temp address management
  - name: emails
    description: Email management
  - name: webhook
    description: Webhook endpoints for receiving emails
  - name: stats
    description: Statistics and summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /subdomains:
    get:
      operationId: listSubdomains
      tags: [subdomains]
      summary: List all subdomains
      responses:
        "200":
          description: List of subdomains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subdomain"
    post:
      operationId: createSubdomain
      tags: [subdomains]
      summary: Add a new subdomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubdomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "409":
          description: Subdomain already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subdomains/{id}:
    get:
      operationId: getSubdomain
      tags: [subdomains]
      summary: Get a subdomain by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subdomain found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subdomain"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSubdomain
      tags: [subdomains]
      summary: Delete a subdomain and all its emails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subdomains/{id}/emails:
    get:
      operationId: listEmailsBySubdomain
      tags: [emails]
      summary: List emails for a subdomain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
        "404":
          description: Subdomain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List all generated temp addresses
      responses:
        "200":
          description: List of generated addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TempAddress"
    post:
      operationId: generateAddress
      tags: [addresses]
      summary: Generate a new random temp email address at a subdomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Address generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempAddress"
        "404":
          description: Subdomain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /addresses/{id}:
    get:
      operationId: getAddress
      tags: [addresses]
      summary: Get a generated address by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Address found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempAddress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete a generated address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /addresses/{id}/emails:
    get:
      operationId: listEmailsByAddress
      tags: [addresses]
      summary: List all emails received at a specific generated address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of emails for this address
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
        "404":
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails:
    get:
      operationId: listRecentEmails
      tags: [emails]
      summary: List recent emails across all subdomains
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Recent emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"

  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get an email by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEmail
      tags: [emails]
      summary: Delete an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails/{id}/read:
    patch:
      operationId: markEmailRead
      tags: [emails]
      summary: Mark an email as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /webhook/email:
    post:
      operationId: receiveEmailWebhook
      tags: [webhook]
      summary: Webhook endpoint to receive incoming emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEmailPayload"
      responses:
        "200":
          description: Email received and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: No matching subdomain or address found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get overall stats summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Subdomain:
      type: object
      required: [id, name, createdAt, emailCount, unreadCount]
      properties:
        id:
          type: integer
        name:
          type: string
        webhookSecret:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        emailCount:
          type: integer
        unreadCount:
          type: integer

    SubdomainInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 253
        webhookSecret:
          type: string

    TempAddress:
      type: object
      required: [id, localPart, domain, fullAddress, subdomainId, subdomainName, createdAt, emailCount, unreadCount]
      properties:
        id:
          type: integer
        localPart:
          type: string
          description: The part before the @ symbol
        domain:
          type: string
          description: The domain part (matches a registered subdomain name)
        fullAddress:
          type: string
          description: Full email address (localPart@domain)
        subdomainId:
          type: integer
        subdomainName:
          type: string
        createdAt:
          type: string
          format: date-time
        emailCount:
          type: integer
        unreadCount:
          type: integer

    AddressInput:
      type: object
      required: [subdomainId]
      properties:
        subdomainId:
          type: integer
          description: Which subdomain domain to generate an address for
        localPart:
          type: string
          description: Optional custom local part; auto-generated if omitted

    Email:
      type: object
      required: [id, subdomainId, subdomainName, fromAddress, toAddress, subject, bodyText, receivedAt, isRead]
      properties:
        id:
          type: integer
        subdomainId:
          type: integer
        subdomainName:
          type: string
        addressId:
          type: ["integer", "null"]
          description: The generated address this email was matched to, if any
        fromAddress:
          type: string
        toAddress:
          type: string
        subject:
          type: string
        bodyText:
          type: string
        bodyHtml:
          type: ["string", "null"]
        receivedAt:
          type: string
          format: date-time
        isRead:
          type: boolean

    WebhookEmailPayload:
      type: object
      required: [from, to, subject, bodyText]
      properties:
        from:
          type: string
        to:
          type: string
        subject:
          type: string
        bodyText:
          type: string
        bodyHtml:
          type: string

    StatsSummary:
      type: object
      required: [totalSubdomains, totalEmails, unreadEmails, emailsToday, totalAddresses]
      properties:
        totalSubdomains:
          type: integer
        totalAddresses:
          type: integer
        totalEmails:
          type: integer
        unreadEmails:
          type: integer
        emailsToday:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string