# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - temporary email service with custom subdomains and generated addresses
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  /** Which subdomain domain to generate an address for */
  subdomainId: number;
  /** Optional custom local part; auto-generated if omitted */
  localPart?: string;
}